separation between log and non-log viscosity equations
